# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/workflows/{id}": {
      "get": {
        "description": "Includes its steps and webhook token.",
        "operationId": "Workflows_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Workflow id",
            "schema": {
              "example": "cm0x8b1f40001abcdlkj2h3g4",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found, or not owned by the caller",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded — retry after the window resets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "JWT": []
          }
        ],
        "summary": "Get a workflow",
        "tags": [
          "Workflows"
        ]
      },
      "patch": {
        "description": "Partial update. Sending `steps` replaces the entire step list (past executions keep their history). Changing `cronExpression` or `timezone` reschedules an active workflow immediately.",
        "operationId": "Workflows_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Workflow id",
            "schema": {
              "example": "cm0x8b1f40001abcdlkj2h3g4",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkflowDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed, or the body contains no fields to update",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found, or not owned by the caller",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded — retry after the window resets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "503": {
            "description": "A dependency (PostgreSQL or Redis) is temporarily unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "JWT": []
          }
        ],
        "summary": "Update a workflow",
        "tags": [
          "Workflows"
        ]
      },
      "delete": {
        "description": "Removes its schedule and all of its execution history.",
        "operationId": "Workflows_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Workflow id",
            "schema": {
              "example": "cm0x8b1f40001abcdlkj2h3g4",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Workflow deleted"
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found, or not owned by the caller",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded — retry after the window resets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "503": {
            "description": "A dependency (PostgreSQL or Redis) is temporarily unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "JWT": []
          }
        ],
        "summary": "Delete a workflow",
        "tags": [
          "Workflows"
        ]
      }
    },
...
    "/api/v1/workflows/{workflowId}/executions": {
      "post": {
        "description": "Creates a `PENDING` execution and queues it for a worker, returning **202 Accepted** immediately. Poll `statusUrl` to follow progress. Paused workflows can still be run manually.",
        "operationId": "WorkflowExecutions_trigger",
        "parameters": [
          {
            "name": "workflowId",
            "required": true,
            "in": "path",
            "description": "Workflow id",
            "schema": {
              "example": "cm0x8b1f40001abcdlkj2h3g4",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionAcceptedDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found, or not owned by the caller",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded — retry after the window resets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "503": {
            "description": "Job queue (Redis) unavailable — the execution is recorded as FAILED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "JWT": []
          }
        ],
        "summary": "Run a workflow now",
        "tags": [
          "Executions"
        ]
      },
      "get": {
        "description": "Newest first, filterable by status and trigger.",
        "operationId": "WorkflowExecutions_list",
        "parameters": [
          {
            "name": "workflowId",
            "required": true,
            "in": "path",
            "description": "Workflow id",
            "schema": {
              "example": "cm0x8b1f40001abcdlkj2h3g4",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "1-based page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Only return executions with this status",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "RUNNING",
                "SUCCESS",
                "FAILED",
                "SKIPPED"
              ]
            }
          },
          {
            "name": "triggerType",
            "required": false,
            "in": "query",
            "description": "Only return executions started this way",
            "schema": {
              "type": "string",
              "enum": [
                "MANUAL",
                "CRON",
                "WEBHOOK"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedExecutionsResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found, or not owned by the caller",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded — retry after the window resets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "JWT": []
          }
        ],
        "summary": "List executions of a workflow",
        "tags": [
          "Executions"
        ]
      }
    },
...
  },
  "info": {
    "title": "CronFlow API",
    "description": "CronFlow runs multi-step workflows — HTTP calls and delays — on a **cron schedule**, from a **webhook**, or **on demand**.\n\n### Quick start\n1. `POST /api/v1/auth/register` and copy the `accessToken`\n2. Click **Authorize**, paste the token into **JWT**\n3. `POST /api/v1/workflows` to create a workflow (see the request examples)\n4. `POST /api/v1/workflows/{workflowId}/executions` to run it now, then poll `GET /api/v1/executions/{id}`\n\n### Authentication\n| Method | Header | Obtain from |\n| --- | --- | --- |\n| JWT | `Authorization: Bearer <accessToken>` | `POST /api/v1/auth/login` |\n| API key | `x-api-key: cf_…` | `POST /api/v1/api-keys` |\n\n### Errors\nEvery error response uses the same envelope (`ErrorResponseDto`). Its `requestId` matches the\n`X-Request-Id` response header and the server logs.\n\n### Rate limiting\nRequests are limited per client IP (100/minute by default; 10/minute for register and login).\nExceeding the limit returns `429 Too Many Requests`.",
    "version": "3.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access token from POST /api/v1/auth/login"
      },
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key from POST /api/v1/api-keys"
      }
    },
    "schemas": {
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Bad Request",
            "description": "HTTP reason phrase"
          },
          "message": {
            "description": "A single message, or one message per failed validation rule",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": [
              "cronExpression must be a valid 5-field cron expression (minute hour day-of-month month day-of-week)"
            ]
          },
          "path": {
            "type": "string",
            "example": "/api/v1/workflows"
          },
          "method": {
            "type": "string",
            "example": "POST"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-09-14T10:00:00.000Z",
            "format": "date-time"
          },
          "requestId": {
            "type": "string",
            "example": "5f0e7c1a-3b8e-4d7a-9c2f-1e6b8a4d2c10",
            "description": "Correlates this response with server logs (also sent as the X-Request-Id header)"
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "path",
          "method",
          "timestamp"
        ]
      },
...
      "HttpStepConfigDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "example": "https://httpbin.org/post",
            "description": "Absolute http(s) URL to call"
          },
          "method": {
            "type": "string",
            "enum": [
              "GET",
              "POST",
              "PUT",
              "PATCH",
              "DELETE",
              "HEAD"
            ],
            "default": "GET"
          },
          "headers": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "example": {
              "Authorization": "Bearer <token>"
            },
            "description": "Request headers"
          },
          "body": {
            "type": "object",
            "description": "JSON body, sent with POST, PUT, PATCH and DELETE requests",
            "example": {
              "event": "nightly-sync"
            }
          }
        },
        "required": [
          "url"
        ]
      },
      "DelayStepConfigDto": {
        "type": "object",
        "properties": {
          "duration": {
            "type": "number",
            "example": 2000,
            "minimum": 1,
            "maximum": 300000,
            "description": "How long to pause, in milliseconds"
          }
        },
        "required": [
          "duration"
        ]
      },
      "WorkflowStepDto": {
        "type": "object",
        "properties": {
          "stepOrder": {
            "type": "number",
            "example": 1,
            "minimum": 1,
            "maximum": 1000,
            "description": "Execution order. Must be unique within the workflow."
          },
          "type": {
            "type": "string",
            "enum": [
              "HTTP",
              "DELAY"
            ],
            "example": "HTTP"
          },
          "config": {
            "description": "Shape depends on `type` — HTTP: `{ url, method?, headers?, body? }` · DELAY: `{ duration }`",
            "oneOf": [
              {
                "$ref": "#/components/schemas/HttpStepConfigDto"
              },
              {
                "$ref": "#/components/schemas/DelayStepConfigDto"
              }
            ]
          },
          "retryCount": {
            "type": "number",
            "minimum": 0,
            "maximum": 10,
            "default": 3,
            "description": "HTTP steps: retries after a transient failure (exponential backoff starting at 1s)"
          },
          "timeout": {
            "type": "number",
            "minimum": 1000,
            "maximum": 300000,
            "default": 30000,
            "description": "HTTP steps: per-attempt timeout in milliseconds"
          }
        },
        "required": [
          "stepOrder",
          "type",
          "config"
        ]
      },
...
      "WorkflowStepResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cm0xa1b2c0002abcd3e4f5g6h"
          },
          "stepOrder": {
            "type": "number",
            "example": 1
          },
          "type": {
            "type": "string",
            "enum": [
              "HTTP",
              "DELAY"
            ],
            "example": "HTTP"
          },
          "config": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/HttpStepConfigDto"
              },
              {
                "$ref": "#/components/schemas/DelayStepConfigDto"
              }
            ],
            "example": {
              "url": "https://httpbin.org/post",
              "method": "POST"
            }
          },
          "retryCount": {
            "type": "number",
            "example": 3
          },
          "timeout": {
            "type": "number",
            "example": 30000,
            "description": "Per-attempt timeout in milliseconds"
          }
        },
        "required": [
          "id",
          "stepOrder",
          "type",
          "config",
          "retryCount",
          "timeout"
        ]
      },
      "WorkflowResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cm0x8b1f40001abcdlkj2h3g4"
          },
          "name": {
            "type": "string",
            "example": "Nightly data sync"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Warms the reporting cache, waits, then triggers the export"
          },
          "cronExpression": {
            "type": "string",
            "example": "0 2 * * *"
          },
          "timezone": {
            "type": "string",
            "example": "UTC"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "PAUSED"
            ],
            "example": "ACTIVE"
          },
          "nextRunAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": "2026-09-15T02:00:00.000Z",
            "description": "Next scheduled run; null while the workflow is paused"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-09-14T10:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-09-14T10:00:00.000Z"
          },
          "webhookToken": {
            "type": "string",
            "example": "k3Jd9sQ2mVx7LpA0bR4tYw8eZc1uN6hG",
            "description": "Secret for `POST /api/v1/webhooks/{token}/trigger`. Treat it like a password."
          },
          "steps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkflowStepResponseDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "cronExpression",
          "timezone",
          "status",
          "nextRunAt",
          "createdAt",
          "updatedAt",
          "webhookToken",
          "steps"
        ]
      },
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          },
          "totalItems": {
            "type": "number",
            "example": 42,
            "description": "Total items matching the filters"
          },
          "totalPages": {
            "type": "number",
            "example": 3
          },
          "hasNextPage": {
            "type": "boolean",
            "example": true
          },
          "hasPreviousPage": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "page",
          "limit",
          "totalItems",
          "totalPages",
          "hasNextPage",
          "hasPreviousPage"
        ]
      },
...
      "UpdateWorkflowDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Nightly data sync",
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "example": "Warms the reporting cache, waits, then triggers the export",
            "maxLength": 500
          },
          "cronExpression": {
            "type": "string",
            "example": "0 2 * * *",
            "description": "Standard 5-field cron expression: minute hour day-of-month month day-of-week"
          },
          "timezone": {
            "type": "string",
            "example": "Europe/London",
            "default": "UTC",
            "description": "IANA timezone the cron expression is evaluated in"
          },
          "steps": {
            "minItems": 1,
            "maxItems": 20,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkflowStepDto"
            }
          }
        }
      },
...
      "ExecutionAcceptedDto": {
        "type": "object",
        "properties": {
          "executionId": {
            "type": "string",
            "example": "cm0xb2c3d0002abcd4e5f6g7h"
          },
          "workflowId": {
            "type": "string",
            "example": "cm0x8b1f40001abcdlkj2h3g4"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "RUNNING",
              "SUCCESS",
              "FAILED",
              "SKIPPED"
            ],
            "example": "PENDING"
          },
          "triggerType": {
            "type": "string",
            "enum": [
              "MANUAL",
              "CRON",
              "WEBHOOK"
            ],
            "example": "MANUAL"
          },
          "statusUrl": {
            "type": "string",
            "example": "/api/v1/executions/cm0xb2c3d0002abcd4e5f6g7h",
            "description": "Poll this to follow progress"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-09-14T10:00:00.000Z"
          }
        },
        "required": [
          "executionId",
          "workflowId",
          "status",
          "triggerType",
          "statusUrl",
          "createdAt"
        ]
      },
      "ExecutionSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cm0xb2c3d0002abcd4e5f6g7h"
          },
          "workflowId": {
            "type": "string",
            "example": "cm0x8b1f40001abcdlkj2h3g4"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "RUNNING",
              "SUCCESS",
              "FAILED",
              "SKIPPED"
            ],
            "example": "SUCCESS"
          },
          "triggerType": {
            "type": "string",
            "enum": [
              "MANUAL",
              "CRON",
              "WEBHOOK"
            ],
            "example": "MANUAL"
          },
          "errorMessage": {
            "type": "string",
            "nullable": true,
            "example": null,
            "description": "Why the execution failed, if it did"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": "2026-09-14T10:00:00.100Z"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": "2026-09-14T10:00:02.600Z"
          },
          "durationMs": {
            "type": "number",
            "nullable": true,
            "example": 2500
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-09-14T10:00:00.000Z"
          }
        },
        "required": [
          "id",
          "workflowId",
          "status",
          "triggerType",
          "errorMessage",
          "startedAt",
          "completedAt",
          "durationMs",
          "createdAt"
        ]
      },
      "PaginatedExecutionsResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExecutionSummaryDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
    }
  }
}